Break logki into modules
